Enable customization of TinyMCE init options
tinymce_init_callback and tinymce_setup_callback can only be used effectively to set events or do something after tinyMCE has been initialized, to change init options tinymce currently needs to be initialized, destroyed and reinitialized with new settings.

This PR:
Exports default settings options
Moves initialization to rich_text_box.blade file
Checks if custom options are set in BREAD details to merge
Enables different settings for each textarea

Closes #4551
