+++ docs/bread/formfields/tinymce.md
@@ -1,0 +1,13 @@
+# TinyMCE
+
+If you want to customize TinyMCE init configuration Options you can merge your custom options as follow:
+
+```php
+{
+    "tinymceOptions" : {
+        "name": "value"
+    }
+}
+```
+
+For all possible configuration Options please refer to the [TinyMCE documentation](https://www.tinymce.com/docs/api/tinymce/tinymce.editor/).
