+++ resources/views/formfields/rich_text_box.blade.php
@@ -1,3 +1,17 @@
 <textarea class="form-control richTextBox" name="{{ $row->field }}" id="richtext{{ $row->field }}">
     {{ old($row->field, $dataTypeContent->{$row->field} ?? '') }}
 </textarea>
+
+@push('javascript')
+    <script>
+        $(document).ready(function() {
+            var additionalConfig = {
+                selector: 'textarea.richTextBox[name="{{ $row->field }}"]',
+            }
+
+            $.extend(additionalConfig, {!! json_encode($options->tinymceOptions ?? '{}') !!})
+
+            tinymce.init(window.voyagerTinyMCE.getConfig(additionalConfig));
+        });
+    </script>
+@endpush
