+++ resources/views/settings/index.blade.php
@@ -483,6 +483,9 @@
                 $(this).closest('form').attr('action', $(this).attr('href'));
                 $(this).closest('form').submit();
             });
+
+            // Initiliaze rich text editor
+            tinymce.init(window.voyagerTinyMCE.getConfig());
         });
     </script>
     <script type="text/javascript">
